+++ HISTORY.rst
@@ -16,7 +16,14 @@ History
 [To do]
 -------
 * Add on/off switch characters
-* Update module publication
+
+?.?.? (2026-07-11)
+------------------
+* added details option to transliterate, storing in transliterator object
+* modernized dynamic imports (importlib.util)
+* removed Marshmallow warnings
+* Add transliterate_with_details() function returning tuple
+* switched to ruff
 
 1.2.4 (2023-10-15)
 ------------------